backend/pyproject.toml
[project]
name = "dms-backend"
version = "0.1.0"
description = "Document Management System Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.12",
    "celery[redis]>=5.4.0",
    "aiobotocore>=2.15.0",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "Pillow>=11.0.0",
    "python-telegram-bot>=21.7",
    "python-magic>=0.4.27",
    "aiofiles>=24.1.0",
    "pydantic[email]>=2.10.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=2.0.0",
    "pyzbar>=0.1.9",
    "aiohttp>=3.9.0",
    "redis>=5.0.0",
    "authlib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
